made supervise data a json serializable object
