--- lib/src/logic/data/SuperviseData.dart
@@ -1,13 +1,0 @@
-class SuperviseData {
-  final String data;
-  bool flagged;
-  bool deleted;
-  String methodName;
-  String display1;
-  String display2;
-  String name;
-  int teamNum;
-
-  SuperviseData(
-      {required this.data, required this.flagged, required this.deleted, required this.methodName, required this.display1, required this.display2, required this.name, required this.teamNum});
-}
