+++ lib/src/views/ui/sushi_supervise/edit.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:localstore/localstore.dart';
 import 'package:sushi_scouts/src/logic/Constants.dart';
-import 'package:sushi_scouts/src/logic/data/SuperviseData.dart';
+import 'package:sushi_scouts/src/logic/models/supervise_data.dart';
 import 'package:sushi_scouts/src/logic/data/config_file_reader.dart';
 import 'package:sushi_scouts/src/logic/helpers/color/hex_color.dart';
 import 'package:sushi_scouts/src/views/util/opacityfilter.dart';
@@ -40,7 +40,7 @@ class _EditState extends State<Edit> {
   Future<void> refreshData() async {
     final newData =
         await Localstore.instance.collection(SUPERVISE_DATABASE_NAME).get();
-
+    print(newData);
     if (newData != null) {
       setState(() {
         for (var name in newData.keys) {
@@ -57,6 +57,7 @@ class _EditState extends State<Edit> {
             )
           });
         }
+        print(data);
       });
     }
   }
